+++ src/server/entities/adjustedWidgetLayout.ts
