+++ src/server/service/transformWidgetConfigService.test.ts
@@ -2,8 +2,8 @@
  * @jest-environment node
  */
 
-import { AdjustedWidgetLayout } from "../entities/adjustedWidgetConfig";
-import { UserWidgetLayout } from "../entities/userWidgetConfig";
+import { AdjustedWidgetLayout } from "../entities/adjustedWidgetLayout";
+import { UserWidgetLayout } from "../entities/userWidgetLayout";
 import transformWidgetLayout from "./transformWidgetConfigService";
 
 describe("transformWidgetConfigService", () => {
